Change mapper return types from Integer to Long
Updated BaseMapper and related interfaces to use Long instead of Integer for insert, update, and count methods to support larger result values. Fixed query parameter references in XML and Java builder for date filtering. Updated test to use new query parameters and validate count method.

+++ sql2codeInitDemo/src/main/java/com/slilio/sql2codeInitDemo/mappers/BaseMapper.java
@@ -5,20 +5,20 @@ import org.apache.ibatis.annotations.Param;
 
 public interface BaseMapper<T, P> {
   /** insert：插入 */
-  Integer insert(@Param("bean") T t);
+  Long insert(@Param("bean") T t);
 
   /** insertOrUpdate：插入或者更新 */
-  Integer insertOrUpdate(@Param("bean") T t);
+  Long insertOrUpdate(@Param("bean") T t);
 
   /** insertBatch：批量插入 */
-  Integer updateBatch(@Param("list") List<T> list);
+  Long updateBatch(@Param("list") List<T> list);
 
   /** insertOrUpdateBatch：批量插入或更新 */
-  Integer insertOrUpdateBatch(@Param("list") List<T> list);
+  Long insertOrUpdateBatch(@Param("list") List<T> list);
 
   /** selectList：根据参数查询集合 */
   List<T> selectList(@Param("query") P p);
 
   /** selectCount */
-  Integer selectCount(@Param("query") P p);
+  Long selectCount(@Param("query") P p);
 }
